+++ server/app.py
@@ -12,20 +12,30 @@ from server.messages import MessageData
 
 import flask
 
-# App
-app = flask.Flask(__name__)
 
-# this one is global because we only have one event loop that receives messages
-message_data = MessageData(req_port=5556, sub_port=5558)
+def build_app(req_port=5556, sub_port=5558):
+    """App is already global and existing"""
+    global app
+    global message_data
 
-# Register the blueprints
-for blueprint in blueprints.get_blueprints():
-    url_prefix = '/' + blueprint.name
-    app.register_blueprint(blueprint, url_prefix=url_prefix)
+    # App
+    app = flask.Flask(__name__)
 
-# Setup logging
-loghelper.setup_logging(logfile_name='server.log')
+    # this one is global because we only have one event loop that receives messages
+    message_data = MessageData(req_port=5556, sub_port=5558)
+
+    # Register the blueprints
+    for blueprint in blueprints.get_blueprints():
+        url_prefix = '/' + blueprint.name
+        app.register_blueprint(blueprint, url_prefix=url_prefix)
+
+    # Setup logging
+    loghelper.setup_logging(logfile_name='server.log')
+    print("request port: %d (server should process requests on this port)" % req_port)
+    print("subscription port: %d (server should publish on this port)" % sub_port)
+
+    return app
 
 # Main
 def run():
-    app.run(host='0.0.0.0', debug=True)
+    build_app(req_port=5556, sub_port=5558).run(host='0.0.0.0', debug=True)