+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog of threedi-wms
 
 - Added option maxpoints to gettimeseries: throw away points till you have the max number of points you have :-)
 
+- Add find link for netCDF4 1.0.4 to buildout.cfg.
+
 
 0.7 (2014-02-05)
 ----------------